moved error report
