+++ octizys_common/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod identifier;
 pub mod logic_path;
+pub mod report;
 pub mod span;
 
 pub mod equivalence;
